Добавлены ресурсы. Настроены стили

+++ app/src/main/java/com/ishabaev/weather/data/Main.java
@@ -14,7 +14,7 @@ public class Main {
     private Double pressure;
 
     @SerializedName("humidity")
-    private Integer humidity;
+    private Double humidity;
 
     @SerializedName("temp_min")
     private Double tempMin;
@@ -38,11 +38,11 @@ public class Main {
         this.pressure = pressure;
     }
 
-    public Integer getHumidity() {
+    public Double getHumidity() {
         return humidity;
     }
 
-    public void setHumidity(Integer humidity) {
+    public void setHumidity(Double humidity) {
         this.humidity = humidity;
     }
